requirements.txt
nose
sphinx
duecredit
pandas==2.2.3
networkx==3.4.2
matplotlib
matplotlib-inline==0.1.7
umap-learn==0.5.7
scipy>=1.2.1
scikit-learn>=0.19.2
